delete ContactPreferenceService, add UserService getPreferredContact
